+++ PeMem/PeMem.vcxproj
@@ -141,15 +141,10 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Address.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Process.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Address.h" />
-    <ClInclude Include="Process.h" />
-    <ClInclude Include="Time.h" />
-    <ClInclude Include="WorkingDirectory.h" />
+    <ClInclude Include="PeMem.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
