+++ PeMem/PeMem.vcxproj.filters
@@ -1,14 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="Address.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Process.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Address.h" />
-    <ClInclude Include="Process.h" />
-    <ClInclude Include="Time.h" />
-    <ClInclude Include="WorkingDirectory.h" />
+    <ClInclude Include="PeMem.h" />
   </ItemGroup>
 </Project>
